Fix leaks with directory passed as command
If a directory was passed as a command it would be passed to execve()
which then returned and was interpreted as an error. Thus, the string
containing the name of the binary was not freed.

Now, the string containing the executables path will be checked, if
this is only a directory it will not be executed and the appropriate
error message will be outputted.