env: distinguish executed binaries
Binaries were executed through env as if the shell executed them
directly. This resulted to even execute the built-ins. Moreover, the
error messages were ouputted as if the execution came from the shell.

Looking directly for the binaries without a previous check for built-ins
solves this. Also, error messages have been updated accordingly.